refactor: ChessBoard생성을 PieceFactory를 통해 하도록 변경
