+++ README.md
@@ -18,7 +18,6 @@
     - [x] 체스판 위치를 벗어날 수 없다.
     - [x] 아군 말이 있는 곳은 이동할 수 없다.
 
-
 - [x] 말의 초기 위치를 지정해준다.
 - [x] 말의 초기 색을 지정해준다.
 
@@ -62,3 +61,13 @@
 ## 체스게임
 
 - [x] 게임 시작과 종료를 제어한다.
+
+- status입력이 들어오면 현재의 각 진영의 점수를 출력하고, 이긴 진영을 출력한다.
+- 각 piece에는 점수가 있다.
+  - Queen - 9점
+  - Rook - 5점
+  - Bishop - 3점
+  - Knight - 2.5점
+  - Pawn - 1점  --> 같은 세로줄(File)의 같은 색(Team)의 Pawn은 0.5점
+- 팀 별로 각각 점수를 계산해야한다.
+- King이 잡히면 게임이 종료되어야 한다(체크메이트). -> 매번 양팀의 King이 살아있는지 확인할 것인가?
